requirements.txt
flask
pandas
psycopg2
flask-cors
gunicorn
python-dotenv
requests
scikit-learn
numpy